+++ R/check.R
@@ -62,17 +62,14 @@
 #' x <- rfast99(params = params, n = 200, q = q, q.arg = q.arg, rep = 20)
 #'
 #' time <- seq(from = 0.25, to = 12.25, by = 0.5)
-#' y <- solve_fun(x, model = FFPK, time = time, vars = "output")
+#' out <- solve_fun(x, model = FFPK, time = time, vars = "output")
 #'
-#'
-#' tell2(x,y) # Link decoupling simulation result
-#' x
 #'
 #' # Check results of sensitivity measures
-#' check(x)
-#' plot(x)
-#' heat_check(x)
-#' heat_check(x, index = "CI")
+#' check(out)
+#' plot(out)
+#' heat_check(out)
+#' heat_check(out, index = "CI")
 #'
 #' @seealso \code{\link{tell2}}
 #'
@@ -205,7 +202,7 @@ heat_check <- function(x,
 
   X <- X %>% filter(order %in% !!(order)) %>% filter_(~variable %in% vars) %>% filter(time %in% times)
 
-  if(length(times) < 10){
+  if(length(times) < 16){
     X$time <- as.factor(X$time)
   }
 
@@ -223,7 +220,7 @@ heat_check <- function(x,
       scale_fill_gradient(low = "white", high = "red", limits = c(-0.05,1.05))
   }
 
-  if(length(times) < 10){
+  if(length(times) < 16){
     p <- p + scale_x_discrete(expand=c(0,0))
   } else p <- p + scale_x_continuous(expand=c(0,0))