+++ R/solve_mcsim.R
@@ -8,7 +8,7 @@
 #'
 #' @param mName a string giving the name of the model or C file (without extension).
 #' @param x a list of storing information in the defined sensitivity function.
-#' @param n a numeric to define the sample number.
+#' @param monte_carlo a numeric to define the sample number in Monte Carlo simulaton.
 #' @param dist a vector of distribution names corresponding to \code{<distribution-name>} in MCSim.
 #' @param q.arg a list of shape parameters in the sampling distribution (\code{dist}).
 #' @param infile.name a character to assign the name of input file.
@@ -20,7 +20,8 @@
 #' @param condition a character to set the specific parameter value in the input file.
 #' @param rtol an argument passed to the integrator (default 1e-6).
 #' @param atol an argument passed to the integrator (default 1e-9).
-#' @param generate.infile a logical value to automatically generate the input file, .
+#' @param generate.infile a logical value to automatically generate the input file.
+#' @param tell a logical value to automatically combine the result y to decoupling simulation x.
 #'
 #' @importFrom utils write.table
 #' @importFrom data.table fread
@@ -63,42 +64,57 @@
 solve_mcsim <- function(x, mName,
                         infile.name = NULL,
                         outfile.name = NULL,
-                        n = NULL,
                         setpoint.name = NULL,
                         params = NULL,
                         vars  = NULL,
                         time  = NULL,
                         condition  = NULL,
-                        generate.infile = T){
+                        generate.infile = T,
+                        tell = T,
+                        rtol = 1e-6, atol = 1e-6,
+                        monte_carlo = NULL, dist = NULL, q.arg = NULL){
 
+  message(paste0("Starting time: ", Sys.time()))
   if(is.null(infile.name)) infile.name <- "input.in"
-  if(is.null(outfile.name)) outfile.name <- "output.csv"
+  if(is.null(outfile.name)) outfile.name <- "sim.out"
 
   if(generate.infile == T){
-    generate_infile(infile.name = infile.name,
-                    outfile.name = outfile.name,
-                    params = params,
-                    vars = vars,
-                    time = time,
-                    condition = condition)
+    if (is.null(monte_carlo)){
+      generate_infile(infile.name = infile.name,
+                      outfile.name = outfile.name,
+                      params = params,
+                      vars = vars,
+                      time = time,
+                      condition = condition)
+    } else { # must be Monte Carlo
+      generate_infile(infile.name = infile.name,
+                      outfile.name = outfile.name,
+                      params = params,
+                      vars = vars,
+                      time = time,
+                      condition = condition,
+                      rtol = rtol, atol = atol,
+                      monte_carlo = monte_carlo, dist = dist, q.arg = q.arg)
+    }
   }
 
-  if(is.null(condition) && is.null(setpoint.name) && is.null(n)){
+  if(is.null(condition) && is.null(setpoint.name) && is.null(monte_carlo)){
     stop("Please assign the setpoint.name (parameter matrix defined in input file)")
   }
 
   if(!is.null(condition)){
-    setpoint.data <- "setpoint.dat"
+    setpoint.data <- "setpoint.out"
   } else setpoint.data <- setpoint.name
 
   mcsim. <- paste0("mcsim.", mName)
-  #if(file.exists(mcsim.) == F){
-  #  stop(paste0("The ", "mcsim.", mName, " doesn't exist."))
-  #}
+  mcsim. <- ifelse(file.exists(mcsim.) ,mcsim., paste0(mcsim., ".exe")) # Design for MCSim under R
+  if(file.exists(mcsim.) == F){
+    stop(paste0("The ", mcsim., " doesn't exist."))
+  }
 
   #
-  if (is.numeric(n)){
-    n.sample <- n
+  if (is.numeric(monte_carlo)){
+    n.sample <- monte_carlo
   } else if (!is.null(x$s)){
     n.sample <- length(x$s)
   }
@@ -113,18 +129,17 @@ solve_mcsim <- function(x, mName,
   n.vars <- length(vars)
 
   #
-  if (is.null(n)){ # Remember to define n if used external parameter matrix
+  if (is.null(monte_carlo)){ # Remember to define n if used external parameter matrix
     X <- cbind(1, apply(x$a, 3L, c))
     write.table(X, file=setpoint.data, row.names=F, sep="\t")
   }
 
   if(file.exists(mcsim.) == T){
-    system(paste0("./mcsim.", mName, " ", infile.name))
-  } else if (file.exists(paste0(mcsim., ".model.exe")) == T) {
-    system(paste0("./mcsim.", mName, ".model.exe ", infile.name))
+    message(paste0("Execute: ", "./", mcsim., " ", infile.name))
+    system(paste0("./", mcsim., " ", infile.name))
   }
 
-  if (is.null(n)){rm(X)}
+  if (is.null(monte_carlo)){rm(X)}
 
   invisible(gc()); # clean memory
 
@@ -152,12 +167,24 @@ solve_mcsim <- function(x, mName,
 
   invisible(gc()); # clean memory
 
-  dimnames(y)[[3]] <- time
-  dimnames(y)[[4]] <- vars
+  if (length(time) > 1) {
+    dimnames(y)[[3]] <- time
+    dimnames(y)[[4]] <- vars
+  } else {
+    dimnames(y)[[3]] <- list(time)
+    dimnames(y)[[4]] <- list(vars)
+  }
+
+  if (is.null(monte_carlo) && tell == T){
+    tell2(x, y)
+  }
 
   #file.remove(setpoint.data)
+  message(paste0("Ending time: ", Sys.time()))
 
-  return(y)
+  if (is.null(monte_carlo) && tell == T){
+    return(x)
+  } else return(y)
 }
 
 #' @export
@@ -165,12 +192,12 @@ solve_mcsim <- function(x, mName,
 generate_infile <- function(infile.name = NULL,
                             outfile.name = NULL,
                             params, vars, time,
-                            condition, rtol = 1e-6, atol = 1e-9,
-                            n = NULL, dist = NULL, q.arg = NULL){ # Monte Carlo
+                            condition, rtol = 1e-6, atol = 1e-6,
+                            monte_carlo = NULL, dist = NULL, q.arg = NULL){ # Monte Carlo
 
   if(is.null(infile.name)) infile.name <- "input.in"
-  if(is.null(outfile.name)) outfile.name <- "output.csv"
-  setpoint.data <- "setpoint.dat"
+  if(is.null(outfile.name)) outfile.name <- "sim.out"
+  setpoint.data <- "setpoint.out"
 
   #if(file.exists(paste0(infile.name)) == T){
   #  if(menu(c("Yes", "No"),
@@ -185,14 +212,14 @@ generate_infile <- function(infile.name = NULL,
       "----------------------------------------", "\n\n",
       file = infile.name, sep = "")
   cat("Integrate (Lsodes, ", rtol, ", ", atol, " , 1);", "\n\n", file=infile.name, append=TRUE, sep="")
-  if(is.null(n)){
+  if(is.null(monte_carlo)){
     cat("SetPoints (", "\n",
         "\"", outfile.name, "\", \n\"", setpoint.data, "\",\n",
         "0, ", "\n",
         paste(params, collapse = ", "),");\n\n",
         file = infile.name, append = TRUE, sep = "")
   } else {
-    cat("MonteCarlo (", "\"", outfile.name, "\"", ",", n , ",", sample(1:99999, 1), ");\n\n",
+    cat("MonteCarlo (", "\"", outfile.name, "\"", ",", monte_carlo , ",", sample(1:99999, 1), ");\n\n",
         file = infile.name, append = TRUE, sep = "")
     for (i in 1 : length(params)){
       cat("Distrib ( ", params[i], ",", dist[i], ",", paste(unlist(q.arg[i]), collapse = ","), ");", "\n",